+++ src/views/MainWindow/MainWindow.tsx
@@ -186,7 +186,7 @@ function MainWindow(props: MainWindowProps) {
         </nav>
         <div className={classes.app}>
           {shouldShowToolbar &&
-            <WebToolbar theme={props.theme} title="X-NoTES" handleAddNoteButton={handleAddNote}
+            <WebToolbar theme={props.theme} title="Axion Notes" handleAddNoteButton={handleAddNote}
                         isDeleteAllButtonDisabled={isDeleteAllButtonDisabled}
                         handleDeleteAllNotesButton={() => setDeleteAllNotesDialogOpen(true)}
                         handleSettingsButton={() => setSettingsDialogOpen(true)} />